Move DecomposedFoodDesription for cohesion.
