+++ src/features/suggestions/parser/DecomposedFoodDescription.test.ts
