+++ src/features/suggestions/parser/DecomposedFoodDescription.ts
